refactor CLI Hugging Face username helper
