+++ src/openenv/cli/commands/fork.py
@@ -13,7 +13,7 @@ from typing import Annotated
 import typer
 from huggingface_hub import HfApi, login, whoami
 
-from .._cli_utils import console
+from .._cli_utils import _extract_hf_username, console
 
 app = typer.Typer(
     help="Fork (duplicate) an OpenEnv environment on Hugging Face to your account"
@@ -37,19 +37,7 @@ def _parse_key_value(s: str) -> tuple[str, str]:
 def _ensure_hf_authenticated() -> str:
     """Ensure user is authenticated with Hugging Face. Returns username."""
     try:
-        user_info = whoami()
-        if isinstance(user_info, dict):
-            username = (
-                user_info.get("name")
-                or user_info.get("fullname")
-                or user_info.get("username")
-            )
-        else:
-            username = (
-                getattr(user_info, "name", None)
-                or getattr(user_info, "fullname", None)
-                or getattr(user_info, "username", None)
-            )
+        username = _extract_hf_username(whoami())
         if not username:
             raise ValueError("Could not extract username from whoami response")
         console.print(f"[bold green]✓[/bold green] Authenticated as: {username}")
@@ -60,19 +48,7 @@ def _ensure_hf_authenticated() -> str:
         )
         try:
             login()
-            user_info = whoami()
-            if isinstance(user_info, dict):
-                username = (
-                    user_info.get("name")
-                    or user_info.get("fullname")
-                    or user_info.get("username")
-                )
-            else:
-                username = (
-                    getattr(user_info, "name", None)
-                    or getattr(user_info, "fullname", None)
-                    or getattr(user_info, "username", None)
-                )
+            username = _extract_hf_username(whoami())
             if not username:
                 raise ValueError("Could not extract username from whoami response")
             console.print(f"[bold green]✓[/bold green] Authenticated as: {username}")
