+++ src/openenv/cli/commands/push.py
@@ -19,7 +19,7 @@ import typer
 import yaml
 from huggingface_hub import HfApi, login, whoami
 
-from .._cli_utils import console, validate_env_structure
+from .._cli_utils import _extract_hf_username, console, validate_env_structure
 
 app = typer.Typer(help="Push an OpenEnv environment to Hugging Face Spaces")
 
@@ -241,22 +241,6 @@ def _validate_openenv_directory(directory: Path) -> tuple[str, dict]:
         raise typer.BadParameter("openenv.yaml must contain a 'name' field")
 
     return env_name, manifest
-
-
-def _extract_hf_username(user_info: object) -> str | None:
-    """Extract a username from the supported Hugging Face whoami shapes."""
-    if isinstance(user_info, dict):
-        return (
-            user_info.get("name")
-            or user_info.get("fullname")
-            or user_info.get("username")
-        )
-
-    return (
-        getattr(user_info, "name", None)
-        or getattr(user_info, "fullname", None)
-        or getattr(user_info, "username", None)
-    )
 
 
 def _get_hf_username() -> str:
