+++ tests/test_cli/test_fork.py
@@ -6,6 +6,7 @@
 
 """Tests for the openenv fork command."""
 
+from types import SimpleNamespace
 from unittest.mock import MagicMock, patch
 
 from openenv.cli.__main__ import app
@@ -50,6 +51,27 @@ def test_fork_calls_duplicate_space_with_from_id() -> None:
         assert call_kwargs["private"] is False
         # HF API requires hardware; default to free cpu-basic when not specified
         assert call_kwargs["hardware"] == "cpu-basic"
+
+
+def test_fork_authenticates_object_whoami_response() -> None:
+    """Test that fork accepts object-shaped whoami responses."""
+    with (
+        patch("openenv.cli.commands.fork.whoami") as mock_whoami,
+        patch("openenv.cli.commands.fork.login") as mock_login,
+        patch("openenv.cli.commands.fork.HfApi") as mock_hf_api_class,
+    ):
+        mock_whoami.return_value = SimpleNamespace(username="testuser")
+        mock_api = MagicMock()
+        mock_api.duplicate_space.return_value = (
+            "https://huggingface.co/spaces/testuser/source-space"
+        )
+        mock_hf_api_class.return_value = mock_api
+
+        result = runner.invoke(app, ["fork", "owner/source-space"])
+
+        assert result.exit_code == 0
+        mock_login.assert_not_called()
+        mock_api.duplicate_space.assert_called_once()
 
 
 def test_fork_passes_private_and_to_id() -> None:
